Process schedule on the server
