+++ package.json
@@ -19,6 +19,7 @@
     "@remix-run/cloudflare": "^1.8.2",
     "@remix-run/cloudflare-pages": "^1.8.2",
     "@remix-run/react": "^1.8.2",
+    "@remix-run/server-runtime": "^1.8.2",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "clsx": "^1.2.1",
     "cross-env": "^7.0.3",
